Different hole shapes are possible in the same structure
Now scattering on holes and on pillars are separated, and different arrays hole_positions and pillar_positions exist. Also, along with hole_positions the is a list hole_shapes in which shapes of the holes ('circle', 'rectangle') are written. So there can be holes of different shape in the same structure, plus pillars everywhere.